Consolidate load cell utilities
