+++ README.md
@@ -6,10 +6,10 @@ cell through an AL2205 IO-Link hub connected to an AL1342 IO master.
 ## Contents
 
 ```
-IO_master.py            # Modbus communication with the AL1342
-devices/AL2205_Hub.py   # Access analog values from the AL2205 hub
-devices/LoadCell_LCM300.py # Convert analog values to force readings
-calibrateLC.py          # GUI to monitor five load cells simultaneously
+IO_master.py               # Modbus communication with the AL1342
+AL2205_Hub.py              # Access analog values from the AL2205 hub
+LoadCell_LCM300.py         # Convert analog values to force readings
+test_read_five_load_cells.py # Command line utility to read up to five load cells
 ```
 
 ## Installation
@@ -27,8 +27,8 @@ is on channel X1.0.
 
 ```python
 from IO_master import IO_master
-from devices.AL2205_Hub import AL2205Hub
-from devices.LoadCell_LCM300 import LoadCellLCM300
+from AL2205_Hub import AL2205Hub
+from LoadCell_LCM300 import LoadCellLCM300
 
 io = IO_master()                          # Defaults to IP 192.168.100.1
 hub = AL2205Hub(io, port_number=1)        # Hub connected to port 1
@@ -37,10 +37,10 @@ cell = LoadCellLCM300(hub, x1_index=0)    # Load cell on channel X1.0
 print(cell.read_force("N"))              # Read force in newtons
 ```
 
-To launch a window showing live readings from five load cells:
+To read up to five load cells from the command line:
 
 ```
-python calibrateLC.py 192.168.100.1
+python test_read_five_load_cells.py 192.168.100.1
 ```
 
 ## License
